backend/requirements.txt
fastapi
uvicorn
pydantic
python-Levenshtein